fix: renaming to widget_size
