+++ lib/widgets.dart
@@ -1,2 +1,2 @@
-export './src/utils/widget_size_enum.dart';
+export './src/utils/widget_size.dart';
 export './src/widgets/avatar.dart';
